+++ tomcat/src/main/java/nextstep/jwp/controller/HelloController.java
@@ -1,5 +1,6 @@
-package org.apache.coyote.http11.controller;
+package nextstep.jwp.controller;
 
+import org.apache.catalina.controller.AbstractController;
 import org.apache.coyote.http11.message.ContentType;
 import org.apache.coyote.http11.message.HttpStatus;
 import org.apache.coyote.http11.message.request.HttpRequest;
