+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -1,9 +1,9 @@
-package org.apache.coyote.http11.controller;
+package nextstep.jwp.controller;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UnauthorizedException;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.exception.FileNotReadableException;
+import org.apache.catalina.controller.AbstractController;
 import org.apache.coyote.http11.message.ContentType;
 import org.apache.coyote.http11.message.HttpStatus;
 import org.apache.coyote.http11.message.request.HttpRequest;
@@ -14,7 +14,7 @@ import org.apache.coyote.http11.session.SessionManager;
 public class LoginController extends AbstractController {
 
     @Override
-    protected void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) throws FileNotReadableException {
+    protected void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) {
         final Session session = httpRequest.getSession(false);
         if (session != null) {
             httpResponse.setHttpStatus(HttpStatus.FOUND);
@@ -27,7 +27,7 @@ public class LoginController extends AbstractController {
     }
 
     @Override
-    protected void doPost(final HttpRequest httpRequest, final HttpResponse httpResponse) throws FileNotReadableException {
+    protected void doPost(final HttpRequest httpRequest, final HttpResponse httpResponse) {
         try {
             final User user = login(httpRequest);
             httpResponse.setHttpStatus(HttpStatus.FOUND);
