+++ tomcat/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,9 +1,9 @@
-package org.apache.coyote.http11.controller;
+package nextstep.jwp.controller;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.MemberExistException;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.exception.FileNotReadableException;
+import org.apache.catalina.controller.AbstractController;
 import org.apache.coyote.http11.message.ContentType;
 import org.apache.coyote.http11.message.HttpStatus;
 import org.apache.coyote.http11.message.request.HttpRequest;
@@ -12,14 +12,14 @@ import org.apache.coyote.http11.message.response.HttpResponse;
 public class RegisterController extends AbstractController {
 
     @Override
-    protected void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) throws FileNotReadableException {
+    protected void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) {
         final String fileContent = fileReader.readStaticFile("/register.html");
         httpResponse.setHttpStatus(HttpStatus.OK);
         httpResponse.setBody(fileContent, ContentType.findResponseContentTypeFromRequest(httpRequest));
     }
 
     @Override
-    protected void doPost(final HttpRequest httpRequest, final HttpResponse httpResponse) throws FileNotReadableException {
+    protected void doPost(final HttpRequest httpRequest, final HttpResponse httpResponse) {
         try {
             register(httpRequest);
         } catch (MemberExistException e) {
