+++ tomcat/src/main/java/org/apache/catalina/UrlHandlerMapping.java
@@ -1,11 +1,12 @@
-package org.apache.coyote;
+package org.apache.catalina;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
-import org.apache.coyote.http11.controller.FileController;
-import org.apache.coyote.http11.controller.HelloController;
-import org.apache.coyote.http11.controller.LoginController;
-import org.apache.coyote.http11.controller.RegisterController;
+import org.apache.catalina.controller.Controller;
+import org.apache.catalina.controller.FileController;
+import nextstep.jwp.controller.HelloController;
+import nextstep.jwp.controller.LoginController;
+import nextstep.jwp.controller.RegisterController;
 import org.apache.coyote.http11.message.request.HttpRequest;
 
 public class UrlHandlerMapping {
