+++ tomcat/src/main/java/org/apache/catalina/exception/FileNotReadableException.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.exception;
+package org.apache.catalina.exception;
 
 public class FileNotReadableException extends RuntimeException {
