+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -4,9 +4,9 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
-import org.apache.coyote.Controller;
+import org.apache.catalina.controller.Controller;
 import org.apache.coyote.Processor;
-import org.apache.coyote.UrlHandlerMapping;
+import org.apache.catalina.UrlHandlerMapping;
 import org.apache.coyote.http11.exception.UnsupportedContentTypeException;
 import org.apache.coyote.http11.message.HttpStatus;
 import org.apache.coyote.http11.message.request.HttpRequest;
