+++ tomcat/src/test/java/org/apache/coyote/FileReaderTest.java
@@ -2,6 +2,7 @@ package org.apache.coyote;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.catalina.FileReader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
