+++ tomcat/src/test/java/org/apache/coyote/http11/controller/HelloControllerTest.java
@@ -7,7 +7,8 @@ import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import org.apache.coyote.Controller;
+import nextstep.jwp.controller.HelloController;
+import org.apache.catalina.controller.Controller;
 import org.apache.coyote.http11.message.request.HttpRequest;
 import org.apache.coyote.http11.message.response.HttpResponse;
 import org.junit.jupiter.api.DisplayName;
