+++ tomcat/src/test/java/org/apache/coyote/http11/controller/LoginControllerTest.java
@@ -7,8 +7,9 @@ import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.UUID;
-import org.apache.coyote.Controller;
-import org.apache.coyote.FileReader;
+import nextstep.jwp.controller.LoginController;
+import org.apache.catalina.controller.Controller;
+import org.apache.catalina.FileReader;
 import org.apache.coyote.http11.message.request.HttpRequest;
 import org.apache.coyote.http11.message.response.HttpResponse;
 import org.apache.coyote.http11.session.Session;
