+++ tomcat/src/test/java/org/apache/coyote/http11/controller/RegisterControllerTest.java
@@ -6,8 +6,9 @@ import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import org.apache.coyote.Controller;
-import org.apache.coyote.FileReader;
+import nextstep.jwp.controller.RegisterController;
+import org.apache.catalina.controller.Controller;
+import org.apache.catalina.FileReader;
 import org.apache.coyote.http11.message.request.HttpRequest;
 import org.apache.coyote.http11.message.response.HttpResponse;
 import org.junit.jupiter.api.DisplayName;
